+++ nextjs-dashboard/app/dashboard/(overview)/loading.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import DataboardSkeleton from '@/app/ui/skeletons'
+
+function loading() {
+  return (
+    <DataboardSkeleton />
+  )
+}
+
+export default loading
